glib: Move `type_data()` and `get_type()` to new `unsafe trait`
Since this is implemented by `#[object_subclass]`/`#[object_interface]`
it mostly doesn't require changes in dependent crates.

Fixes https://github.com/gtk-rs/gtk-rs/issues/215.
