+++ glib-macros/src/object_interface_attribute.rs
@@ -47,8 +47,26 @@ pub fn impl_object_interface(input: &syn::ItemImpl) -> TokenStream {
         #(#attrs)*
         #unsafety impl#generics #trait_path for #self_ty {
             #prerequisites_opt
-            #crate_ident::object_interface_internal!();
             #(#items)*
+        }
+
+        unsafe impl #crate_ident::subclass::interface::ObjectInterfaceType for #self_ty {
+            fn get_type() -> #crate_ident::Type {
+                static ONCE: std::sync::Once = std::sync::Once::new();
+                static mut TYPE: #crate_ident::Type = #crate_ident::Type::INVALID;
+
+                ONCE.call_once(|| {
+                    let type_ = #crate_ident::subclass::register_interface::<Self>();
+                    unsafe {
+                        TYPE = type_;
+                    }
+                });
+
+                unsafe {
+                    assert!(TYPE.is_valid());
+                    TYPE
+                }
+            }
         }
     }
 }
