+++ glib/src/subclass/interface.rs
@@ -78,27 +78,14 @@ macro_rules! prerequisite_list_trait_inner(
 
 prerequisite_list_trait!(A, B, C, D, E, F, G, H, I, J, K, L, M, N, O, P, Q, R, S);
 
-#[macro_export]
-#[doc(hidden)]
-macro_rules! object_interface_internal {
-    () => {
-        fn get_type() -> $crate::Type {
-            static ONCE: std::sync::Once = std::sync::Once::new();
-            static mut TYPE: $crate::Type = $crate::Type::INVALID;
-
-            ONCE.call_once(|| {
-                let type_ = $crate::subclass::register_interface::<Self>();
-                unsafe {
-                    TYPE = type_;
-                }
-            });
-
-            unsafe {
-                assert!(TYPE.is_valid());
-                TYPE
-            }
-        }
-    };
+/// Type methods required for an [`ObjectInterface`] implementation.
+///
+/// This is usually generated by the [`#[object_interface]`](crate::object_interface) attribute macro.
+pub unsafe trait ObjectInterfaceType {
+    /// Returns the `glib::Type` ID of the interface.
+    ///
+    /// This will register the type with the type system on the first call.
+    fn get_type() -> Type;
 }
 
 /// The central trait for defining a `GObject` interface.
@@ -113,7 +100,7 @@ macro_rules! object_interface_internal {
 /// with the type system.
 ///
 /// [`register_interface`]: fn.register_interface.html
-pub unsafe trait ObjectInterface: Sized + 'static {
+pub unsafe trait ObjectInterface: ObjectInterfaceType + Sized + 'static {
     /// `GObject` type name.
     ///
     /// This must be unique in the whole process.
@@ -124,14 +111,6 @@ pub unsafe trait ObjectInterface: Sized + 'static {
     /// Any implementer of the interface must be a subclass of the prerequisites or implement them
     /// in case of interfaces.
     type Prerequisites: PrerequisiteList;
-
-    /// Returns the `glib::Type` ID of the interface.
-    ///
-    /// This will register the type with the type system on the first call and is usually generated
-    /// by the [`object_interface!`] macro.
-    ///
-    /// [`object_interface!`]: ../../macro.object_interface.html
-    fn get_type() -> Type;
 
     /// Additional type initialization.
     ///
