+++ glib/src/subclass/mod.rs
@@ -250,11 +250,11 @@ pub mod signal;
 pub mod prelude {
     //! Prelude that re-exports all important traits from this crate.
     pub use super::boxed::BoxedType;
-    pub use super::interface::{ObjectInterface, ObjectInterfaceExt};
+    pub use super::interface::{ObjectInterface, ObjectInterfaceExt, ObjectInterfaceType};
     pub use super::object::{ObjectClassSubclassExt, ObjectImpl, ObjectImplExt};
     pub use super::types::{
         ClassStruct, InstanceStruct, IsImplementable, IsSubclassable, ObjectSubclass,
-        ObjectSubclassExt,
+        ObjectSubclassExt, ObjectSubclassType,
     };
 }
