+++ glib/src/subclass/types.rs
@@ -305,37 +305,17 @@ impl TypeData {
     }
 }
 
-#[macro_export]
-#[doc(hidden)]
-macro_rules! object_subclass_internal {
-    () => {
-        fn type_data() -> std::ptr::NonNull<$crate::subclass::TypeData> {
-            static mut DATA: $crate::subclass::TypeData = $crate::subclass::TypeData {
-                type_: $crate::Type::INVALID,
-                parent_class: std::ptr::null_mut(),
-                class_data: None,
-                private_offset: 0,
-            };
+/// Type methods required for an [`ObjectSubclass`] implementation.
+///
+/// This is usually generated by the [`#[object_subclass]`](crate::object_subclass) attribute macro.
+pub unsafe trait ObjectSubclassType {
+    /// Storage for the type-specific data used during registration.
+    fn type_data() -> ptr::NonNull<TypeData>;
 
-            unsafe { std::ptr::NonNull::new_unchecked(&mut DATA) }
-        }
-
-        fn get_type() -> $crate::Type {
-            static ONCE: std::sync::Once = std::sync::Once::new();
-
-            ONCE.call_once(|| {
-                $crate::subclass::register_type::<Self>();
-            });
-
-            unsafe {
-                let data = Self::type_data();
-                let type_ = data.as_ref().get_type();
-                assert!(type_.is_valid());
-
-                type_
-            }
-        }
-    };
+    /// Returns the `glib::Type` ID of the subclass.
+    ///
+    /// This will register the type with the type system on the first call.
+    fn get_type() -> Type;
 }
 
 /// The central trait for subclassing a `GObject` type.
@@ -348,7 +328,7 @@ macro_rules! object_subclass_internal {
 /// with the type system.
 ///
 /// [`register_type`]: fn.register_type.html
-pub trait ObjectSubclass: Sized + 'static {
+pub trait ObjectSubclass: ObjectSubclassType + Sized + 'static {
     /// `GObject` type name.
     ///
     /// This must be unique in the whole process.
@@ -399,21 +379,6 @@ pub trait ObjectSubclass: Sized + 'static {
     // TODO: Should default to simple::ClassStruct<Self> once associated
     // type defaults are stabilized https://github.com/rust-lang/rust/issues/29661
     type Class: ClassStruct<Type = Self>;
-
-    /// Storage for the type-specific data used during registration.
-    ///
-    /// This is usually generated by the [`object_subclass!`] macro.
-    ///
-    /// [`object_subclass!`]: ../../macro.object_subclass.html
-    fn type_data() -> ptr::NonNull<TypeData>;
-
-    /// Returns the `glib::Type` ID of the subclass.
-    ///
-    /// This will register the type with the type system on the first call and is usually generated
-    /// by the [`object_subclass!`] macro.
-    ///
-    /// [`object_subclass!`]: ../../macro.object_subclass.html
-    fn get_type() -> Type;
 
     /// Additional type initialization.
     ///
